+++ Dump.h
@@ -68,11 +68,11 @@ private:
 	string filename;
 	string srcPort;
 	string dstPort;
-	int sentPacketCount;
-	int sentBytesCount;
-	int recvPacketCount;
-	int recvBytesCount;
-	int ackCount;
+	uint64_t sentPacketCount;
+	uint64_t sentBytesCount;
+	uint64_t recvPacketCount;
+	uint64_t recvBytesCount;
+	uint64_t ackCount;
 	uint32_t max_payload_size;
 	map<ConnectionMapKey*, Connection*, ConnectionKeyComparator> conns;
